backend/pyproject.toml
[tool.poetry]
name = "burnup-backend"
version = "0.1.0"
description = "FastAPI backend for burnup chart"
authors = ["Burnup Chart <dev@local>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.5"
uvicorn = {version = "0.30.6", extras = ["standard"]}
numpy = "^2.4.0"
pandas = "^2.3.3"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
bcrypt = "^5.0.0"
python-multipart = "^0.0.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
